Remove constants file
